added saving and importing optimized loss_function

+++ starter_5/utils_starter_5.py
@@ -5,7 +5,7 @@ import matplotlib.pyplot as plt
 import math
 import sys
 import os
-sys.path.append(os.getcwd())
+sys.path.append(os.getcwd()) #to access current working directory files easily
 from image import Image
 
 class Utils_starter_5:
@@ -21,19 +21,59 @@ class Utils_starter_5:
         return
 
     def get_pix_at_translated(self, x : int, y : int, p : list):
+        """_summary_
+
+        Args:
+            x (int): _description_
+            y (int): _description_
+            p (list): _description_
+
+        Returns:
+            _type_: _description_
+        """
         n,m = self._img2.data.shape
         int_px = math.floor(p[0]) #beware indexation
         int_py = math.floor(p[1]) #beware indexation
+
+        #Unoptimised for numpy
+        """
         if 1<=x-int_px<n and 1<=y-int_py<m:
             float_px = p[0] - int_px #TODO probably bugs here with indexation
             float_py = p[1] - int_py
-            return    (1-float_px)*(1-float_py) *self._img2.data[x-int_px][y-int_py] \
-                    + float_px*(1-float_py)     *self._img2.data[x-int_px-1][y-int_py] \
-                    + (1-float_px)*float_py     *self._img2.data[x -int_px][y-int_py-1] \
-                    + float_px*float_py         *self._img2.data[x -int_px-1][y-int_py-1] #TODO probably bugs here with indexation
-
+            return    (1-float_px)*( (1-float_py) *self._img2.data[x-int_px][y-int_py] \
+                                    + float_py    *self._img2.data[x -int_px][y-int_py-1]) \
+                    + float_px   * ( (1-float_py) *self._img2.data[x-int_px-1][y-int_py] \
+                                    + float_py    *self._img2.data[x -int_px-1][y-int_py-1]) #TODO probably bugs here with indexation
         else:
             return 1 #white padding TODO check this condition when calculating loss
+        """
+
+        #Optimised
+        float_px = p[0] - int_px
+        float_py = p[1] - int_py
+        #optimised conditional
+        condition_matrix = np.logical_and(np.logical_and(1+int_px<=x, x<n+int_px) ,
+                                        np.logical_and(1+int_py<=y, y<m+int_py))
+        # condition_matrix = np.zeros((n,m)).astype(bool)
+        # condition_matrix[n//2,m//2] = True
+
+        xcondition = x.copy()
+        ycondition = y.copy()
+        xcondition[np.logical_not(condition_matrix)] = int_px +1 # used to nullify the index in the np.where below
+        ycondition[np.logical_not(condition_matrix)] = int_py +1
+
+        return np.where(condition_matrix,
+                        
+                        (1-float_px)*( (1-float_py) *self._img2.data[xcondition-int_px,   ycondition-int_py] \
+                                        + float_py  *self._img2.data[xcondition -int_px,  ycondition-int_py-1]) \
+                        + float_px  *( (1-float_py) *self._img2.data[xcondition-int_px-1, ycondition-int_py] \
+                                        + float_py  *self._img2.data[xcondition -int_px-1,ycondition-int_py-1]),
+                         1,)
+        
+        #1 is white padding TODO check this condition when calculating loss
+
+
+
 
     def loss_function_1(self,**kwargs):#(self, p : list, warp : callable = get_pix_at_translated): #[[int,int,list],float]
         p = 0
@@ -43,7 +83,14 @@ class Utils_starter_5:
                 p = kwargs['p']
             if params == 'warp':
                 warp = kwargs['warp']
-        warped_img2 = np.array([[warp(i,j,p) for j in range(self._img2.data.shape[1])] for i in range(self._img2.data.shape[0])])
+        
+        # Old version
+        # warped_img2 = np.array([[warp(i,j,p) for j in range(self._img2.data.shape[1])] for i in range(self._img2.data.shape[0])])
+
+        # New version optimized
+        i,j = np.meshgrid(np.arange(self._img2.data.shape[0]), np.arange(self._img2.data.shape[1]),indexing="ij")
+        
+        warped_img2 = warp(i,j,p) # uses numpy arrays here
         return np.sum((self._img1.data - warped_img2)**2)
     
     def loss_function_2(self,**kwargs):#(self, p : list, warp : callable = get_pix_at_translated):
@@ -57,6 +104,68 @@ class Utils_starter_5:
         pass #TODO
 
     
+    def make_save_name(self, loss_function:callable) :
+        return  self._img1.name + "_" + self._img2.name + "_" + str(loss_function.__name__) + ".txt"
+
+    def import_plot(self, loss_function:callable) :
+        """Imports loss_function data from named .txt file
+
+        Args:
+            loss_function (callable): loss function used for finding the name of the file
+
+        Returns:
+            np.ndarray: meshgrid x indexes for p parameter (px)
+            np.ndarray: meshgrid y indexes for p parameter (py)
+            np.ndarray: imported data from loss function (loss_grid)
+        """
+
+        save_filename = self.make_save_name(loss_function)
+        
+        if os.path.exists(save_filename) :
+            with open(save_filename, "r") as f :
+                line1 = f.readline().rstrip('\n').split(" ")
+                print(line1)
+                span_x = 0
+                span_y = 0
+                step = 1
+                for i, word in enumerate(line1) :
+                    if word == "span_x" :
+                        span_x = [int(line1[i+1]),
+                                    int(line1[i+2])]
+                    if word == "span_y" :
+                        span_y = [int(line1[i+1]),
+                                    int(line1[i+2])]
+                    if word == "step":
+                        step = int(line1[i+1])
+
+                translate_span_x = span_x[1]
+                translate_span_y = span_y[1]
+
+
+                # Create table and x and y :
+                px, py = np.meshgrid(np.linspace(-translate_span_x,translate_span_x, translate_span_x * 2, endpoint=False).astype(int),
+                                np.linspace(-translate_span_y,translate_span_y, translate_span_y * 2, endpoint=False).astype(int))
+                loss_grid = np.zeros((translate_span_y * 2,translate_span_x * 2) ) # TODO potential error: mixup between x,y and i,j
+            
+
+                for i in range(span_x[0],span_x[1]):
+                    line =  f.readline().rstrip('\n').split(" ")
+                    for j, word in enumerate(line) :
+                        if word!="":
+                            loss_grid[i,j] = float(word)
+
+                return px, py, loss_grid
+        else :
+            compute = input(f"No data for {loss_function.__name__}. Do you want to compute ? (y/n)")
+            while(not compute in ["y","n"]):
+                print("Please enter y or n")
+                compute = input(f"No data for {loss_function.__name__}. Do you want to compute ? (y/n)")
+            if compute == "y":
+                return self.plot_loss(show=False,span="all")
+            else:
+                return np.array([]), np.array([]), np.array([])
+
+            
     def plot_loss(self, **kwargs) : #TODO : vary the ranges and shapes for p, maybe with relation to loss function and image data
         """Function used to greedily calculate all the loss_function returns for a certain range of p
 
@@ -66,7 +175,12 @@ class Utils_starter_5:
                 save : bool, saves as .txt
                 show : bool, plots the function
                 span : "all" for whole span of p
-
+        
+        Returns:
+            np.ndarray: meshgrid x indexes for p parameter (px)
+            np.ndarray: meshgrid y indexes for p parameter (py)
+            np.ndarray: imported data from loss function (loss_grid)
+            
         """
         print("test_plot_loss")
 
@@ -77,12 +191,11 @@ class Utils_starter_5:
         
         n,m = self._img2.data.shape
         
-        ax = plt.figure().add_subplot(projection='3d')
         
         translate_span_x = 1
         translate_span_y = m//3
         
-        
+        #reading kwargs
         for key, value in kwargs.items() :
             if key == "loss_function" :
                 loss_function = value
@@ -97,35 +210,59 @@ class Utils_starter_5:
                 else:
                     raise ValueError("unknown value for span")
 
-        px, py = np.meshgrid(np.linspace(-translate_span_x,translate_span_x, translate_span_x * 2, endpoint=False).astype(int),
-                             np.linspace(-translate_span_y,translate_span_y, translate_span_y * 2, endpoint=False).astype(int))
-        #arbitrary range for p, beware the endpoint is True
+        #check for existing file
+        save_filename = self.make_save_name(loss_function)
+    
+        compute = "y"
+        if os.path.exists(save_filename) :
+                compute = input(f"Data exists for {loss_function.__name__}. Do you want to compute and overwrite ? (y/n)")
+                while(not compute in ["y","n"]):
+                    print("Please enter y or n")
+                    compute = input(f"Data exists for {loss_function.__name__}. Do you want to compute and overwrite ? (y/n)")
 
-        loss_grid = np.zeros(px.shape)
-        for i in range(px.shape[0]):
-            for j in range(px.shape[1]):
-                p = [px[i,j],py[i,j]]
-                print(p)
-                loss_grid[i][j] = loss_function(p = p, warp = self.get_pix_at_translated ) #pass the px and py values
-        
-        print("loss computation done")
+        if compute == "y":
 
-        ax.plot_surface(px,py,loss_grid)
-        if save :
-            save_filename = self._img1.name + "_" + self._img2.name + "_" + str(loss_function.__name__) + ".txt"
-            with open(save_filename, "w") as f :
-                f.write(["span_x", -translate_span_x, translate_span_x,
-                         "span_y ", -translate_span_y, translate_span_y,
-                         "step ", 1].join(" ")) #add step
+            px, py = np.meshgrid(np.linspace(-translate_span_x,translate_span_x, translate_span_x * 2, endpoint=False).astype(int),
+                                np.linspace(-translate_span_y,translate_span_y, translate_span_y * 2, endpoint=False).astype(int))
+            #TODO : check meshgrid indexing
+            #arbitrary range for p, beware the endpoint is True
 
-                for i in range(loss_grid.shape[0]) :
-                    for j in range(loss_grid.shape[1]) :
-                        f.write(str(loss_grid[i,j]) + " ")
+            loss_grid = np.zeros(px.shape)
+            for i in range(px.shape[0]):
+                for j in range(px.shape[1]):
+                    p = [px[i,j],py[i,j]]
+                    print(p)
+                    loss_grid[i][j] = loss_function(p = p, warp = self.get_pix_at_translated ) #pass the px and py values
+            
+            print("loss computation done")
+
+            if save :
+                with open(save_filename, "w") as f :
+                    f.write(" ".join(["span_x", str(-translate_span_x), str(translate_span_x),
+                            "span_y", str(-translate_span_y), str(translate_span_y),
+                            "step", str(1)]) ) #add step
                     f.write("\n")
+                    
+
+                    for i in range(loss_grid.shape[0]) :
+                        f.write(str(loss_grid[i,0]))
+                        for j in range(loss_grid.shape[1]) :
+                            f.write(" " + str(loss_grid[i,j]))
+                        f.write("\n")
 
 
         if show :
+            ax = plt.figure().add_subplot(projection='3d')
+            
+            if compute == "y":
+                ax.plot_surface(px,py,loss_grid)
+            else :
+                px, py, loss_grid = self.import_plot(loss_function)
+                ax.plot_surface(px,py,loss_grid)
+            
             plt.show()
+        
+        return px, py, loss_grid
 
     def greedy_optimization_xy(self, **kwargs) : #TODO : vary the ranges and shapes for p, maybe with relation to loss function and image data
         """greedy brute force strategy to find the optimal value of p_x or of [p_x,p_y]
@@ -204,8 +341,14 @@ class Utils_starter_5:
                     l_list[i][j] = l
             print("The translation in y, x coordinates that minimizes our loss function is ", p_min)
             if plot :
-                p_x, p_y = np.meshgrid(np.arange(- math.ceil(n/2), math.floor(n/2) + 1),np.arange(- math.ceil(m/2), math.floor(m/2) + 1))
                 ax = plt.figure().add_subplot(projection='3d')
+                
+                px_loss, py_loss, loss_data = self.import_plot(loss_function)
+                if px_loss != None:
+                    ax.plot_surface(px_loss,py_loss,loss_data)
+                    plt.show()
+
+                p_x, p_y = np.meshgrid(np.arange(- math.ceil(n/2), math.floor(n/2) + 1),np.arange(- math.ceil(m/2), math.floor(m/2) + 1))
                 ax.plot_surface(p_x,p_y,l_list)
                 plt.show()
 
@@ -316,6 +459,12 @@ class Utils_starter_5:
         print("The translation in y, x coordinates that minimizes our loss function is ", p)
         if plot :
             ax = plt.figure().add_subplot(projection='3d')
+            
+            np.array([])
+
+            px_loss, py_loss, loss_data = self.import_plot(loss_function)
+            if len(px_loss) != 0:
+                ax.plot_surface(px_loss,py_loss,loss_data,alpha=0.7)    
             p_list_np = np.array(p_list).transpose()
             l_list_np = np.array(l_list)
             ax.plot(p_list_np[0],p_list_np[1],l_list_np)
@@ -374,6 +523,7 @@ if __name__ == '__main__' :
     if True:
         utils = Utils_starter_5(Image("images/clean_finger.png"),Image("images/tx_finger.png"))
 
+        # utils.plot_loss()
         utils.plot_loss(span = "all")
 
         p, l_list = utils.coordinate_descent_optimization_xy(plot = True, alpha0 = 0.1, epsilon = 100, epsilon2 = 0.001)
